fix publish to accoutn for mono debug files

+++ src/dotnet/commands/dotnet-build/CompileContext.cs
@@ -426,6 +426,12 @@ namespace Microsoft.DotNet.Tools.Build
                 {
                     Directory.CreateDirectory(directoryName);
                 }
+                
+                if (file.EndsWith(".pdb") && !File.Exists(file))
+                {
+                    continue;
+                }
+                
                 File.Copy(file, destFileName, true);
             }
         }
